Screen Capture API - sort by `created_at`

+++ ScreenCapture/sw.js
@@ -12,20 +12,18 @@ const postMessage = async (message, clientId) => {
 /**
  * IndexedDB - `screen-capture` database
  * */
-const setup = () => {
-  const req = indexedDB.open('screen-capture', 1);
+const req = indexedDB.open('screen-capture', 1);
 
 //  Create store
-  req.onupgradeneeded = e => {
-    const db = e.currentTarget.result;
-    const store = db.createObjectStore('screen-capture', {
-      keyPath: 'id',
-      autoIncrement: true
-    });
+req.onupgradeneeded = e => {
+  const db = e.currentTarget.result;
+  const store = db.createObjectStore('screen-capture', {
+    keyPath: 'id',
+    autoIncrement: true
+  });
 
-    store.createIndex('created_at', 'created_at', { unique: false });
-    store.createIndex('url', 'url', { unique: false });
-  };
+  store.createIndex('created_at', 'created_at', { unique: false });
+  store.createIndex('url', 'url', { unique: false });
 };
 
 // Get Store
@@ -35,29 +33,58 @@ const getStore = (mode = 'readonly') => {
   return new Promise(resolve => {
     request.onsuccess = e => resolve(e.currentTarget.result.transaction('screen-capture', mode).objectStore('screen-capture'));
   });
+};
 
+const getAll_ = async () => {
+  const store = await getStore();
+  const request = await store.getAll();
+  request.onsuccess = e => postMessage(['GET_ALL', e.target.result || []]);
+  request.onerror = () => postMessage(['ERROR', 'GET_ALL ERROR']);
+};
+
+// Get All Sort By `created_at`
+const getAll = async () => {
+  const store = await getStore();
+  const request = await store.index('created_at').openCursor(null, 'prev');
+  const result = [];
+  return new Promise(resolve => {
+    request.onsuccess = async e => {
+      let cursor = e.currentTarget.result;
+
+      while (cursor) {
+        result.push(cursor.value);
+        cursor = await cursor.continue();
+
+        if (!cursor) {
+          await postMessage(['GET_ALL', result]);
+          resolve(result);
+        }
+      }
+    };
+  });
 };
 
 // Add
 const addData = async (val) => {
   const store = await getStore('readwrite');
   const addRequest = store.add(val);
-  addRequest.onsuccess = async () => postMessage(['ADD_DATA', await getAll()]);
-};
-
-const getAll = async () => {
-  const store = await getStore();
-  const request = await store.getAll();
-  request.onsuccess = e => postMessage(['GET_ALL', e.target.result]);
+  addRequest.onsuccess = getAll();
 };
 
 // Delete
 const deleteData = async (ids) => {
   const store = await getStore('readwrite');
+
   ids.forEach(id => {
-    const request = store.delete(Number(id));
-    request.onsuccess = async () => postMessage(['DELETE_DATA', await getAll()]);
-    request.onerror = async () => postMessage(['DELETE_DATA', new Error('CANNOT DELETE')]);
+    const getReq = store.get(Number(id));
+    getReq.onsuccess = (e) => {
+      const isExist = e.currentTarget.result;
+      if (!isExist) return;
+      const request = store.delete(Number(id));
+      request.onsuccess = getAll();
+      request.onerror =  postMessage(['ERROR', 'DELETE_DATA ERROR']);
+    };
+    getReq.onerror = postMessage(['ERROR', 'DELETE_DATA ERROR']);
   });
 };
 
@@ -65,9 +92,7 @@ const deleteData = async (ids) => {
 const clearStore = async () => {
   const store = await getStore('readwrite');
   const request = store.clear();
-  setup();
-  request.onsuccess = async () => postMessage(['CLEAR_STORE', await getAll()]);
-  request.onerror = async () => postMessage(['CLEAR_STORE', new Error('CANNOT DELETE')]);
+  request.onsuccess = postMessage(['CLEAR', []]);
 };
 
 self.addEventListener('message', async (e) => {
@@ -77,5 +102,3 @@ self.addEventListener('message', async (e) => {
   if (type === 'DELETE_DATA') await deleteData(data);
   if (type === 'CLEAR_STORE') await clearStore();
 });
-
-setup();
